grpc: moving IsGRPCEnabled to resolve import cycle issue (#55362)
Moving IsGRPCEnabled from internal/grpc to internal/conf package to
resolve the import cycle issue caused when using IsGRPCEnabled in the
/internal/api/internalapi package.

This issues occurs because internalapi depends on -> internal/grpc
(IsGRPCEnabled) depends on -> internal/conf -> which depends on
internalapi

importing grpc/defaults to internal/api/internalapi/client.go and no
longer got import cycle errors
<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->

(cherry picked from commit 5831097001bb0fa97b2b8a6d26ed428c46bbbc04)
