+++ internal/gitserver/client.go
@@ -31,10 +31,13 @@ import (
 	"github.com/sourcegraph/sourcegraph/internal/actor"
 	"github.com/sourcegraph/sourcegraph/internal/api"
 	"github.com/sourcegraph/sourcegraph/internal/authz"
+
+	"github.com/sourcegraph/sourcegraph/internal/conf"
+	"github.com/sourcegraph/sourcegraph/internal/database"
+
 	"github.com/sourcegraph/sourcegraph/internal/gitserver/gitdomain"
 	"github.com/sourcegraph/sourcegraph/internal/gitserver/protocol"
 	proto "github.com/sourcegraph/sourcegraph/internal/gitserver/v1"
-	internalgrpc "github.com/sourcegraph/sourcegraph/internal/grpc"
 	"github.com/sourcegraph/sourcegraph/internal/grpc/streamio"
 	"github.com/sourcegraph/sourcegraph/internal/httpcli"
 	"github.com/sourcegraph/sourcegraph/internal/lazyregexp"
@@ -605,8 +608,9 @@ func (c *RemoteGitCommand) sendExec(ctx context.Context) (_ io.ReadCloser, err e
 		return nil, err
 	}
 
-	if internalgrpc.IsGRPCEnabled(ctx) {
-		client, err := c.execer.ClientForRepo(repoName)
+ 
+	if conf.IsGRPCEnabled(ctx) {
+		client, err := c.execer.ClientForRepo(ctx, repoName)
 		if err != nil {
 			return nil, err
 		}
@@ -699,11 +703,7 @@ func convertGRPCErrorToGitDomainError(err error) error {
 		return context.Canceled
 	}
 
-	if st.Code() == codes.DeadlineExceeded {
-		return context.DeadlineExceeded
-	}
-
-	for _, detail := range st.Details() {
+ {
 		switch payload := detail.(type) {
 
 		case *proto.ExecStatusPayload:
@@ -762,8 +762,9 @@ func (c *clientImplementor) Search(ctx context.Context, args *protocol.SearchReq
 
 	repoName := protocol.NormalizeRepo(args.Repo)
 
-	if internalgrpc.IsGRPCEnabled(ctx) {
-		client, err := c.ClientForRepo(repoName)
+	if conf.IsGRPCEnabled(ctx) {
+		client, err := c.ClientForRepo(ctx, repoName)
+
 		if err != nil {
 			return false, err
 		}
@@ -883,8 +884,9 @@ func (c *clientImplementor) P4Exec(ctx context.Context, host, user, password str
 		P4Passwd: password,
 		Args:     args,
 	}
-	if internalgrpc.IsGRPCEnabled(ctx) {
-		client, err := c.ClientForRepo("")
+
+	if conf.IsGRPCEnabled(ctx) {
+		client, err := c.ClientForRepo(ctx, "")
 		if err != nil {
 			return nil, nil, err
 		}
@@ -1010,7 +1012,7 @@ func (c *clientImplementor) BatchLog(ctx context.Context, opts BatchLogOptions, 
 
 		var response protocol.BatchLogResponse
 
-		if internalgrpc.IsGRPCEnabled(ctx) {
+		if conf.IsGRPCEnabled(ctx) {
 			client, err := grpcClient.client, grpcClient.dialErr
 			if err != nil {
 				return err
@@ -1167,8 +1169,10 @@ func (c *clientImplementor) RequestRepoUpdate(ctx context.Context, repo api.Repo
 		Since: since,
 	}
 
-	if internalgrpc.IsGRPCEnabled(ctx) {
-		client, err := c.ClientForRepo(repo)
+
+	if conf.IsGRPCEnabled(ctx) {
+		client, err := c.ClientForRepo(ctx, repo)
+
 		if err != nil {
 			return nil, err
 		}
@@ -1205,8 +1209,8 @@ func (c *clientImplementor) RequestRepoUpdate(ctx context.Context, repo api.Repo
 
 // RequestRepoClone requests that the gitserver does an asynchronous clone of the repository.
 func (c *clientImplementor) RequestRepoClone(ctx context.Context, repo api.RepoName) (*protocol.RepoCloneResponse, error) {
-	if internalgrpc.IsGRPCEnabled(ctx) {
-		client, err := c.ClientForRepo(repo)
+	if conf.IsGRPCEnabled(ctx) {
+		client, err := c.ClientForRepo(ctx, repo)
 		if err != nil {
 			return nil, err
 		}
@@ -1258,8 +1262,8 @@ func (c *clientImplementor) IsRepoCloneable(ctx context.Context, repo api.RepoNa
 
 	var resp protocol.IsRepoCloneableResponse
 
-	if internalgrpc.IsGRPCEnabled(ctx) {
-		client, err := c.ClientForRepo(repo)
+	if conf.IsGRPCEnabled(ctx) {
+		client, err := c.ClientForRepo(ctx, repo)
 		if err != nil {
 			return err
 		}
@@ -1336,7 +1340,7 @@ func (e *RepoNotCloneableErr) Error() string {
 func (c *clientImplementor) RepoCloneProgress(ctx context.Context, repos ...api.RepoName) (*protocol.RepoCloneProgressResponse, error) {
 	numPossibleShards := len(c.Addrs())
 
-	if internalgrpc.IsGRPCEnabled(ctx) {
+	if conf.IsGRPCEnabled(ctx) {
 		shards := make(map[proto.GitserverServiceClient]*proto.RepoCloneProgressRequest, (len(repos)/numPossibleShards)*2) // 2x because it may not be a perfect division
 		for _, r := range repos {
 			client, err := c.ClientForRepo(r)
@@ -1458,7 +1462,7 @@ func (c *clientImplementor) ReposStats(ctx context.Context) (map[string]*protoco
 	stats := map[string]*protocol.ReposStats{}
 	var allErr error
 
-	if internalgrpc.IsGRPCEnabled(ctx) {
+	if conf.IsGRPCEnabled(ctx) {
 		for _, addr := range c.clientSource.Addresses() {
 			client, err := addr.GRPCClient()
 			if err != nil {
@@ -1510,8 +1514,8 @@ func (c *clientImplementor) Remove(ctx context.Context, repo api.RepoName) error
 	// In case the repo has already been deleted from the database we need to pass
 	// the old name in order to land on the correct gitserver instance
 	repo = api.UndeletedRepoName(repo)
-	if internalgrpc.IsGRPCEnabled(ctx) {
-		client, err := c.ClientForRepo(repo)
+	if conf.IsGRPCEnabled(ctx) {
+		client, err := c.ClientForRepo(ctx, repo)
 		if err != nil {
 			return err
 		}
@@ -1601,63 +1605,64 @@ func (c *clientImplementor) do(ctx context.Context, repoForTracing api.RepoName,
 }
 
 func (c *clientImplementor) CreateCommitFromPatch(ctx context.Context, req protocol.CreateCommitFromPatchRequest) (*protocol.CreateCommitFromPatchResponse, error) {
-	if internalgrpc.IsGRPCEnabled(ctx) {
-		client, err := c.ClientForRepo(req.Repo)
-		if err != nil {
-			return nil, err
-		}
-		resp, err := client.CreateCommitFromPatchBinary(ctx, req.ToProto())
-		if err != nil {
-			return nil, err
-		}
-
-		var res protocol.CreateCommitFromPatchResponse
-		res.FromProto(resp)
-
-		if resp.GetError() != nil {
-			return &res, errors.New(resp.GetError().String())
-		}
-
-		return &res, nil
-	} else {
-		resp, err := c.httpPost(ctx, req.Repo, "create-commit-from-patch-binary", req)
-		if err != nil {
-			return nil, err
-		}
-		defer resp.Body.Close()
-
-		body, err := io.ReadAll(resp.Body)
-		if err != nil {
-			return nil, errors.Wrap(err, "failed to read response body")
-		}
-		var res protocol.CreateCommitFromPatchResponse
-		if err := json.Unmarshal(body, &res); err != nil {
-			c.logger.Warn("decoding gitserver create-commit-from-patch response", sglog.Error(err))
-			return nil, &url.Error{
-				URL: resp.Request.URL.String(),
-				Op:  "CreateCommitFromPatch",
-				Err: errors.Errorf("CreateCommitFromPatch: http status %d, %s", resp.StatusCode, string(body)),
+	if conf.IsGRPCEnabled(ctx) {
+		client, err := c.ClientForRepo(ctx, req.Repo)
+			if err != nil {
+				return nil, err
 			}
+			resp, err := client.CreateCommitFromPatchBinary(ctx, req.ToProto())
+			if err != nil {
+				return nil, err
+			}
+
+			var res protocol.CreateCommitFromPatchResponse
+			res.FromProto(resp)
+
+			if resp.GetError() != nil {
+				return &res, errors.New(resp.GetError().String())
+			}
+
+			return &res, nil
+		} else {
+			resp, err := c.httpPost(ctx, req.Repo, "create-commit-from-patch-binary", req)
+			if err != nil {
+				return nil, err
+			}
+			defer resp.Body.Close()
+
+			body, err := io.ReadAll(resp.Body)
+			if err != nil {
+				return nil, errors.Wrap(err, "failed to read response body")
+			}
+			var res protocol.CreateCommitFromPatchResponse
+			if err := json.Unmarshal(body, &res); err != nil {
+				c.logger.Warn("decoding gitserver create-commit-from-patch response", sglog.Error(err))
+				return nil, &url.Error{
+					URL: resp.Request.URL.String(),
+					Op:  "CreateCommitFromPatch",
+					Err: errors.Errorf("CreateCommitFromPatch: http status %d, %s", resp.StatusCode, string(body)),
+				}
+			}
+
+			if res.Error != nil {
+				return &res, res.Error
+			}
+			return &res, nil
+		}
+	}
+
+	func (c *clientImplementor) GetObject(ctx context.Context, repo api.RepoName, objectName string) (*gitdomain.GitObject, error) {
+		if ClientMocks.GetObject != nil {
+			return ClientMocks.GetObject(repo, objectName)
 		}
 
-		if res.Error != nil {
-			return &res, res.Error
+		req := protocol.GetObjectRequest{
+			Repo:       repo,
+			ObjectName: objectName,
 		}
-		return &res, nil
-	}
-}
 
-func (c *clientImplementor) GetObject(ctx context.Context, repo api.RepoName, objectName string) (*gitdomain.GitObject, error) {
-	if ClientMocks.GetObject != nil {
-		return ClientMocks.GetObject(repo, objectName)
-	}
-
-	req := protocol.GetObjectRequest{
-		Repo:       repo,
-		ObjectName: objectName,
-	}
-	if internalgrpc.IsGRPCEnabled(ctx) {
-		client, err := c.ClientForRepo(req.Repo)
+	if conf.IsGRPCEnabled(ctx) {
+		client, err := c.ClientForRepo(ctx, req.Repo)
 		if err != nil {
 			return nil, err
 		}
