+++ internal/grpc/BUILD.bazel
@@ -10,7 +10,6 @@ go_library(
     importpath = "github.com/sourcegraph/sourcegraph/internal/grpc",
     visibility = ["//:__subpackages__"],
     deps = [
-        "//internal/conf",
         "@com_github_sourcegraph_log//:log",
         "@org_golang_google_grpc//:go_default_library",
         "@org_golang_google_grpc//codes",
