+++ internal/grpc/grpc.go
@@ -2,17 +2,12 @@
 package grpc
 
 import (
-	"context"
 	"net/http"
-	"os"
-	"strconv"
 	"strings"
 
 	"golang.org/x/net/http2"
 	"golang.org/x/net/http2/h2c"
 	"google.golang.org/grpc"
-
-	"github.com/sourcegraph/sourcegraph/internal/conf"
 )
 
 // MultiplexHandlers takes a gRPC server and a plain HTTP handler and multiplexes the
@@ -38,6 +33,8 @@ const envGRPCEnabled = "SG_FEATURE_FLAG_GRPC"
 func IsGRPCEnabled(ctx context.Context) bool {
 	if val, err := strconv.ParseBool(os.Getenv(envGRPCEnabled)); err == nil {
 		return val
+
 	}
 	return conf.Get().ExperimentalFeatures.EnableGRPC
 }
+
