+++ internal/grpc/grpc_test.go
@@ -2,7 +2,6 @@ package grpc
 
 import (
 	"bytes"
-	"context"
 	"net/http"
 	"net/http/httptest"
 	"testing"
@@ -36,22 +35,5 @@ func TestMultiplexHandlers(t *testing.T) {
 		called = false
 		multiplexedHandler.ServeHTTP(httptest.NewRecorder(), req)
 		require.False(t, called)
-	}
-}
-
-func TestIsGRPCEnabled(t *testing.T) {
-	t.Setenv(envGRPCEnabled, "true")
-	if !IsGRPCEnabled(context.Background()) {
-		t.Fatal("expected grpc to be enabled")
-	}
-
-	t.Setenv(envGRPCEnabled, "false")
-	if IsGRPCEnabled(context.Background()) {
-		t.Fatal("expected grpc to not be enabled")
-	}
-
-	t.Setenv(envGRPCEnabled, "")
-	if IsGRPCEnabled(context.Background()) {
-		t.Fatal("expected grpc to not be enabled")
 	}
 }
