+++ internal/search/backend/BUILD.bazel
@@ -23,7 +23,6 @@ go_library(
         "//internal/conf",
         "//internal/ctags_config",
         "//internal/featureflag",
-        "//internal/grpc",
         "//internal/grpc/defaults",
         "//internal/honey",
         "//internal/httpcli",
