+++ internal/symbols/BUILD.bazel
@@ -10,9 +10,9 @@ go_library(
         "//internal/actor",
         "//internal/api",
         "//internal/authz",
+        "//internal/conf",
         "//internal/conf/conftypes",
         "//internal/endpoint",
-        "//internal/grpc",
         "//internal/grpc/defaults",
         "//internal/httpcli",
         "//internal/limiter",
